+++ modules/hdf5/src/cpp/whoNh5File.cpp
@@ -16,24 +16,28 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include <matio.h>
+#define H5_BUILT_AS_DYNAMIC_LIB
+#include <hdf5.h>
 #include <boost/filesystem.hpp>
 #include <boost/filesystem/path.hpp>
-#include "LoadMatioFile.hpp"
-#include "LoadMatioVariable.hpp"
+#include <boost/container/vector.hpp>
+#include "whoNh5File.hpp"
+#include "h5SaveLoadHelpers.hpp"
+#include "h5LoadVariable.hpp"
 #include "characters_encoding.hpp"
+#include "ClassName.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
 ArrayOf
-LoadMatioFile(Evaluator* eval, const std::wstring& filename, wstringVector names, bool asStruct)
+whoNh5File(Interface* io, const std::wstring& filename, wstringVector names, bool asCell)
 {
     ArrayOf res;
-    boost::filesystem::path mat_filename(filename);
+    boost::filesystem::path nh5_filename(filename);
     bool fileExistPreviously = false;
     try {
-        fileExistPreviously = boost::filesystem::exists(mat_filename)
-            && !boost::filesystem::is_directory(mat_filename);
+        fileExistPreviously = boost::filesystem::exists(nh5_filename)
+            && !boost::filesystem::is_directory(nh5_filename);
     } catch (const boost::filesystem::filesystem_error& e) {
         if (e.code() == boost::system::errc::permission_denied) {
             Error(_W("Permission denied."));
@@ -44,18 +48,17 @@ LoadMatioFile(Evaluator* eval, const std::wstring& filename, wstringVector names
         Error(_W("File does not exist."));
     }
 
-    std::string utf8filename = wstring_to_utf8(filename);
-    mat_t* matfile = Mat_Open(utf8filename.c_str(), MAT_ACC_RDONLY);
-    if (!matfile) {
-        Error(_W("Cannot open .mat file."));
+    hid_t fid
+        = H5Fopen(wstring_to_utf8(nh5_filename.wstring()).c_str(), H5F_ACC_RDONLY, H5P_DEFAULT);
+    if (fid == H5I_INVALID_HID) {
+        Error(_W("Open file failed."));
     }
-
-    stringVector variableNamesInFile;
-    size_t nVars = 0;
-    char** variableNames = Mat_GetDir(matfile, &nVars);
-    for (size_t k = 0; k < nVars; k++) {
-        variableNamesInFile.push_back(variableNames[k]);
+    if (!isNelsonH5File(fid)) {
+        H5Fclose(fid);
+        Error(_W("Invalid file format."));
     }
+    stringVector variableNamesInFile = getVariableNames(fid);
+    H5Fclose(fid);
     stringVector variablesNamesToRead;
     if (names.empty()) {
         variablesNamesToRead = variableNamesInFile;
@@ -65,38 +68,42 @@ LoadMatioFile(Evaluator* eval, const std::wstring& filename, wstringVector names
             if (std::find(variableNamesInFile.begin(), variableNamesInFile.end(), name)
                 != variableNamesInFile.end()) {
                 variablesNamesToRead.push_back(name);
-            } else {
-                std::string msg = _("Variable not found:") + std::string(" ") + name;
-                Warning(msg);
             }
         }
     }
-    ArrayOfVector values;
-    for (std::string name : variablesNamesToRead) {
-        ArrayOf value;
-        matvar_t* matVariable = Mat_VarRead(matfile, name.c_str());
-        if (matVariable == nullptr) {
-            Mat_Close(matfile);
-            std::string msg = _("Cannot read variable:") + std::string(" ") + name;
-            Error(msg);
+    if (asCell) {
+        Dimensions dims(variablesNamesToRead.size(), 1);
+        ArrayOf* elements
+            = (ArrayOf*)ArrayOf::allocateArrayOf(NLS_CELL_ARRAY, dims.getElementCount());
+        indexType k = 0;
+        for (std::string name : variablesNamesToRead) {
+            elements[k] = ArrayOf::characterArrayConstructor(name);
+            k++;
         }
-        bool bSuccess = LoadMatioVariable(matVariable, false, value);
-        Mat_VarFree(matVariable);
-        if (bSuccess) {
-            values.push_back(value);
-        } else {
-            Mat_Close(matfile);
-            std::string msg = _("Cannot read variable:") + std::string(" ") + name;
-            Error(msg);
-        }
-    }
-    Mat_Close(matfile);
-    if (asStruct) {
-        res = ArrayOf::structScalarConstructor(variablesNamesToRead, values);
+        res = ArrayOf(NLS_CELL_ARRAY, dims, elements);
     } else {
-        for (indexType i = 0; i < variablesNamesToRead.size(); i++) {
-            eval->getContext()->getCurrentScope()->insertVariable(
-                variablesNamesToRead[i], values[i]);
+        if (variablesNamesToRead.size()) {
+            size_t ncharmax = io->getTerminalWidth();
+            size_t nbchar = 0;
+            if (!variablesNamesToRead.empty()) {
+                io->outputMessage(_W("Your variables are:") + L"\n\n");
+            }
+            for (auto& k : variablesNamesToRead) {
+                if (nbchar + k.size() < ncharmax) {
+                    io->outputMessage(k);
+                    io->outputMessage(" ");
+                    nbchar = 1 + nbchar + k.size();
+                } else {
+                    nbchar = 0;
+                    io->outputMessage("\n");
+                    io->outputMessage(k);
+                    io->outputMessage(" ");
+                    nbchar = 1 + nbchar + k.size();
+                }
+            }
+            if (!variablesNamesToRead.empty()) {
+                io->outputMessage("\n");
+            }
         }
     }
     return res;